Set views to appropriate height, merge utility files
