+++ Tempo/Controllers/PlayerTableViewController.swift
@@ -77,7 +77,7 @@ class PlayerTableViewController: UIViewController, UITableViewDelegate, UITableV
         super.viewDidLoad()
 		
 		//TableView
-		tableView = UITableView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height - tabBarHeight), style: .plain)
+		tableView = UITableView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height - tabBarHeight - miniPlayerHeight), style: .plain)
 		tableView.delegate = self
 		tableView.dataSource = self
 		playingPostType = .unknown
